Django/manual-instrumentation/todolist_project/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Deprecated==1.2.14
Django==5.1
django-environ==0.11.2
elastic-opentelemetry==0.1.0
googleapis-common-protos==1.63.2
grpcio==1.65.4
idna==3.7
importlib_metadata==7.1.0
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-asyncio==0.46b0
opentelemetry-instrumentation-aws-lambda==0.46b0
opentelemetry-instrumentation-dbapi==0.46b0
opentelemetry-instrumentation-django==0.46b0
opentelemetry-instrumentation-grpc==0.46b0
opentelemetry-instrumentation-logging==0.46b0
opentelemetry-instrumentation-requests==0.46b0
opentelemetry-instrumentation-sqlite3==0.46b0
opentelemetry-instrumentation-system-metrics==0.46b0
opentelemetry-instrumentation-threading==0.46b0
opentelemetry-instrumentation-urllib==0.46b0
opentelemetry-instrumentation-urllib3==0.46b0
opentelemetry-instrumentation-wsgi==0.46b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-test-utils==0.46b0
opentelemetry-util-http==0.46b0
protobuf==4.25.4
psutil==5.9.8
requests==2.32.3
setuptools==72.1.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
uWSGI==2.0.26
wrapt==1.16.0
zipp==3.19.2
